requirements.txt
# =============================================================================
# 高波数 Helmholtz 神经算子 2D 原型 — 依赖清单
# =============================================================================
# 安装: pip install -r requirements.txt

# --- 核心计算 ---
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0

# --- 配置管理 ---
pyyaml>=6.0

# --- 可视化 ---
matplotlib>=3.7.0

# --- 测试 ---
pytest>=7.0.0
pytest-cov>=4.0.0

# --- 开发工具 ---
tqdm>=4.65.0
